Arreglo del maxClones
maxClones= variable que cuenta la cantidad de clones en la escena

+++ Assets/Scenes/scripts/RandomPlacement.cs
@@ -4,9 +4,10 @@ using UnityEngine;
 
 public class RandomPlacement : MonoBehaviour
 {
-    public float offset;
+    public int offset;
     public void SetRandomPosition()
     {
+      
         Vector3 newPosition = new Vector3(Random.Range(0, offset), 9, Random.Range(0, offset));
         transform.position = newPosition;
     }
